Refactor project structures
