+++ src/application/event_handler.rs
@@ -1,6 +1,5 @@
-use crossterm::event::{KeyCode, KeyEvent, KeyEventKind};
-
 use crate::application::{App, InputMode};
+use crossterm::event::{KeyCode, KeyEvent, KeyEventKind};
 
 pub struct EventHandler;
 impl EventHandler {
