+++ src/application/mod.rs
@@ -1,5 +1,9 @@
 pub mod app;
+pub mod event_handler;
+pub mod state;
 pub mod tasks;
 
 pub use app::*;
+pub use event_handler::*;
+pub use state::*;
 pub use tasks::*;
