--- src/database/mod.rs
@@ -1,3 +1,0 @@
-pub mod db;
-
-pub use db::{load, save};
