--- src/event_handler/mod.rs
@@ -1,1 +1,0 @@
-pub mod event_handlers;
