+++ src/lib.rs
@@ -1,5 +1,3 @@
 pub mod application;
-pub mod database;
-pub mod event_handler;
-pub mod state;
-pub mod user_interface;
+pub mod backend;
+pub mod ui;
