--- src/user_interface/mod.rs
@@ -1,5 +1,0 @@
-pub mod terminal;
-pub mod ui;
-
-pub use terminal::Terminal;
-pub use ui::*;
